tests: Update phpunit and enable coverage for all classes
The structure of the Evaluator is that there is a single class
that serves as the main entry point, and everything else is called
internally, so at this point we might as well allow coverage
of all classes from our current tests, which are essentially
integration tests.

This is normally achieved by disabling 'forceCoversAnnotation',
but that is disabled by default and wasn't enabled. The reason
the test coverage report was behaving this way was due to the
presence of one @covers annotation, which I removed.

Also:
* Minor phpunit update.
* Remove redundant variable assignment in test.
* Add 'composer cover' command.
* Add /coverage to gitignore.
* Add visibility to tests.

Change-Id: I9f671dbf5b575799570db307df3fa310fad7976f